Adding professor restriction; updating readme;

+++ src/controller/CourseController.ts
@@ -2,6 +2,7 @@ import { Request, Response } from "express";
 import { getManager } from "typeorm";
 import { Course } from "./../entity/Course";
 import { ApprovedCourse } from "./../entity/ApprovedCourse";
+import { checkUserAllowed } from "./UserController";
 
 var _ = require("underscore")
 
@@ -14,43 +15,60 @@ export async function getAllCourses(request: Request, response: Response) {
 }
 
 export async function deleteCouse(request: Request, response: Response) {
-    const courseRepository = getManager().getRepository(Course);
-    const course = await courseRepository.findOne({
-        where: {
-            id: request.body.id
-        }
-    });
+    var userId = request.header('User-Id');
 
-    await courseRepository.delete(course);
-    response.send(`Course Deleted Id: ${request.body.id}`);
+    if (await checkUserAllowed(userId ? userId : 0)) {
+        const courseRepository = getManager().getRepository(Course);
+        const course = await courseRepository.findOne({
+            where: {
+                id: request.body.id
+            }
+        });
+
+        await courseRepository.delete(course);
+        response.send(`Course Deleted Id: ${request.body.id}`);
+    } else {
+        response.send(`You must be a professor to delete a course.`)
+    }
 }
 
 export async function postCourse(request: Request, response: Response) {
+    var userId = request.header('User-Id');
     var body = request.body;
 
-    const courseRepository = getManager().getRepository(Course);
-    const course = await courseRepository.create(body);
+    if (await checkUserAllowed(userId ? userId : 0)) {
+        const courseRepository = getManager().getRepository(Course);
+        const course = await courseRepository.create(body);
 
-    await courseRepository.save(course);
-    response.send(course);
+        await courseRepository.save(course);
+        response.send(course);
+    } else {
+        response.send(`You must be a professor to post a course.`)
+    }
 }
 
 export async function editCourse(request: Request, response: Response) {
+    var userId = request.header('User-Id');
     var body = request.body;
 
-    const courseRepository = getManager().getRepository(Course);
-    const course = await courseRepository.findOne({
-        where: {
-            id: request.body.id
-        }
-    });
+    if (await checkUserAllowed(userId ? userId : 0)) {
 
-    course.courseName = body.courseName;
-    course.lessons = body.lessons;
-    course.requiredCourses = body.requiredCourses;
+        const courseRepository = getManager().getRepository(Course);
+        const course = await courseRepository.findOne({
+            where: {
+                id: request.body.id
+            }
+        });
 
-    await courseRepository.save(course);
-    response.send(course);
+        course.courseName = body.courseName;
+        course.lessons = body.lessons;
+        course.requiredCourses = body.requiredCourses;
+
+        await courseRepository.save(course);
+        response.send(course);
+    } else {
+        response.send(`You must be a professor to edit a course.`)
+    }
 }
 
 export async function getStudentCourses(request: Request, response: Response) {
@@ -83,17 +101,20 @@ export async function getStudentCourses(request: Request, response: Response) {
         })
     })
 
-    promiseCoursesArray.push(new Promise(async (resolve) => {
-        var coursesWithoutCorrelatives = _.filter(courses, course => {
-            return _.isEmpty(course.requiredCourses);
-        });
-
-        var filteredCourses = _.filter(coursesWithoutCorrelatives, (course) => {
-            return !_.some(approvedCourseList, (approvedCourse) => {
-                return approvedCourse.course.id === course.id
+    promiseCoursesArray.push(new Promise(async (resolve, reject) => {
+        try {
+            var coursesWithoutCorrelatives = _.filter(courses, course => {
+                return _.isEmpty(course.requiredCourses);
+            });
+            var filteredCourses = _.filter(coursesWithoutCorrelatives, (course) => {
+                return !_.some(approvedCourseList, (approvedCourse) => {
+                    return approvedCourse.course.id === course.id
+                })
             })
-        })
-        resolve(filteredCourses);
+            resolve(filteredCourses);
+        } catch (error) {
+            reject(error);
+        }
     }))
 
     Promise.all(promiseCoursesArray).then(async (value) => {
@@ -103,5 +124,7 @@ export async function getStudentCourses(request: Request, response: Response) {
             "availableCourses": availableCourses,
             "allCourses": courses
         })
+    }).catch((error) => {
+        response.send(error)
     })
 }
